Feature/package services

+++ DesafioFinal/src/main/java/com/example/DesafioSprint/Repository/IPackageRepository.java
@@ -3,6 +3,6 @@ package com.example.DesafioSprint.Repository;
 import com.example.DesafioSprint.Entities.TouristicPackage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPackageRepository extends JpaRepository<TouristicPackage,Integer> {
+public interface IPackageRepository extends JpaRepository<TouristicPackage,Long> {
 
 }
